+++ packages/transport-webrtc/test/stream.spec.ts
@@ -4,46 +4,38 @@ import { expect } from 'aegir/chai'
 import length from 'it-length'
 import * as lengthPrefixed from 'it-length-prefixed'
 import { pushable } from 'it-pushable'
+import pDefer from 'p-defer'
 import { Uint8ArrayList } from 'uint8arraylist'
 import { Message } from '../src/pb/message.js'
-import { createStream } from '../src/stream.js'
-
-const mockDataChannel = (opts: { send: (bytes: Uint8Array) => void, bufferedAmount?: number }): RTCDataChannel => {
-  return {
-    readyState: 'open',
-    close: () => {},
-    addEventListener: (_type: string, _listener: () => void) => {},
-    removeEventListener: (_type: string, _listener: () => void) => {},
-    ...opts
-  } as RTCDataChannel
-}
-
-const MAX_MESSAGE_SIZE = 16 * 1024
+import { MAX_BUFFERED_AMOUNT, MAX_MESSAGE_SIZE, PROTOBUF_OVERHEAD, createStream } from '../src/stream.js'
+import { mockDataChannel, receiveFinAck } from './util.js'
 
 describe('Max message size', () => {
   it(`sends messages smaller or equal to ${MAX_MESSAGE_SIZE} bytes in one`, async () => {
     const sent: Uint8ArrayList = new Uint8ArrayList()
-    const data = new Uint8Array(MAX_MESSAGE_SIZE - 5)
+    const data = new Uint8Array(MAX_MESSAGE_SIZE - PROTOBUF_OVERHEAD)
     const p = pushable()
+    const channel = mockDataChannel({
+      send: (bytes) => {
+        sent.append(bytes)
+      }
+    })
 
     // Make sure that the data that ought to be sent will result in a message with exactly MAX_MESSAGE_SIZE
     const messageLengthEncoded = lengthPrefixed.encode.single(Message.encode({ message: data }))
     expect(messageLengthEncoded.length).eq(MAX_MESSAGE_SIZE)
     const webrtcStream = createStream({
-      channel: mockDataChannel({
-        send: (bytes) => {
-          sent.append(bytes)
-        }
-      }),
-      direction: 'outbound'
+      channel,
+      direction: 'outbound',
+      closeTimeout: 1
     })
 
     p.push(data)
     p.end()
+    receiveFinAck(channel)
     await webrtcStream.sink(p)
 
-    // length(message) + message + length(FIN) + FIN
-    expect(length(sent)).to.equal(4)
+    expect(length(sent)).to.equal(6)
 
     for (const buf of sent) {
       expect(buf.byteLength).to.be.lessThanOrEqual(MAX_MESSAGE_SIZE)
@@ -54,22 +46,24 @@ describe('Max message size', () => {
     const sent: Uint8ArrayList = new Uint8ArrayList()
     const data = new Uint8Array(MAX_MESSAGE_SIZE)
     const p = pushable()
+    const channel = mockDataChannel({
+      send: (bytes) => {
+        sent.append(bytes)
+      }
+    })
 
     // Make sure that the data that ought to be sent will result in a message with exactly MAX_MESSAGE_SIZE + 1
     // const messageLengthEncoded = lengthPrefixed.encode.single(Message.encode({ message: data })).subarray()
     // expect(messageLengthEncoded.length).eq(MAX_MESSAGE_SIZE + 1)
 
     const webrtcStream = createStream({
-      channel: mockDataChannel({
-        send: (bytes) => {
-          sent.append(bytes)
-        }
-      }),
+      channel,
       direction: 'outbound'
     })
 
     p.push(data)
     p.end()
+    receiveFinAck(channel)
     await webrtcStream.sink(p)
 
     expect(length(sent)).to.equal(6)
@@ -80,33 +74,32 @@ describe('Max message size', () => {
   })
 
   it('closes the stream if bufferamountlow timeout', async () => {
-    const MAX_BUFFERED_AMOUNT = 16 * 1024 * 1024 + 1
     const timeout = 100
-    let closed = false
-    const webrtcStream = createStream({
-      dataChannelOptions: {
-        bufferedAmountLowEventTimeout: timeout
+    const closed = pDefer()
+    const channel = mockDataChannel({
+      send: () => {
+        throw new Error('Expected to not send')
       },
-      channel: mockDataChannel({
-        send: () => {
-          throw new Error('Expected to not send')
-        },
-        bufferedAmount: MAX_BUFFERED_AMOUNT
-      }),
+      bufferedAmount: MAX_BUFFERED_AMOUNT + 1
+    })
+    const webrtcStream = createStream({
+      bufferedAmountLowEventTimeout: timeout,
+      closeTimeout: 1,
+      channel,
       direction: 'outbound',
       onEnd: () => {
-        closed = true
+        closed.resolve()
       }
     })
 
     const t0 = Date.now()
 
     await expect(webrtcStream.sink([new Uint8Array(1)])).to.eventually.be.rejected
-      .with.property('message', 'Timed out waiting for DataChannel buffer to clear')
+      .with.property('code', 'ERR_BUFFER_CLEAR_TIMEOUT')
     const t1 = Date.now()
     expect(t1 - t0).greaterThan(timeout)
     expect(t1 - t0).lessThan(timeout + 1000) // Some upper bound
-    expect(closed).true()
+    await closed.promise
     expect(webrtcStream.timeline.close).to.be.greaterThan(webrtcStream.timeline.open)
     expect(webrtcStream.timeline.abort).to.be.greaterThan(webrtcStream.timeline.open)
   })