+++ packages/keychain/CHANGELOG.md
@@ -11,6 +11,17 @@
 
 * **dev:** bump aegir from 38.1.8 to 39.0.10 ([#70](https://github.com/libp2p/js-libp2p-keychain/issues/70)) ([4da4a08](https://github.com/libp2p/js-libp2p-keychain/commit/4da4a08b86f436c36e2fae48ecc48817e9b8066f))
 
+### [3.0.4](https://www.github.com/libp2p/js-libp2p/compare/keychain-v3.0.3...keychain-v3.0.4) (2023-09-15)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/crypto bumped from ^2.0.3 to ^2.0.4
+  * devDependencies
+    * @libp2p/peer-id-factory bumped from ^3.0.3 to ^3.0.4
+
 ### [3.0.3](https://www.github.com/libp2p/js-libp2p/compare/keychain-v3.0.2...keychain-v3.0.3) (2023-08-14)